+++ CoreShell.Blazor/CoreShellBlazorServiceCollectionExtensions.cs
@@ -13,6 +13,7 @@ public static class CoreShellBlazorServiceCollectionExtensions
 
         services.TryAddSingleton<ICoreShellContentResolver, BlazorCoreShellContentResolver>();
         services.TryAddSingleton<ICoreShellLayoutResolver, BlazorCoreShellLayoutResolver>();
+        services.TryAddSingleton<ICoreShellNotificationService, EmptyCoreShellNotificationService>();
         services.TryAddSingleton<ICoreShellBlazorPageProjectionService, CoreShellBlazorPageProjectionService>();
         services.TryAddSingleton<
             ICoreShellBlazorSectionOutletProjectionService,